Fixed mkdir bug; some other nitpickings
